.ipynb_checkpoints/requirements-checkpoint.txt
# Backend Dependencies for RFx Agents
# Install: pip install -r requirements.txt

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Strands Agent Framework
strands-agents

# Document Processing
python-docx
PyMuPDF==1.23.26

# JSON & Data Processing
jsonschema>=4.0.0

# Optional: For enhanced features
# requests>=2.31.0
# Pillow>=10.0.0  # For image processing if needed

mcp>=0.9.0  # MCP protocol support